Working authoritative server.

+++ Robotron/Robotron/NetworkComponent.h
@@ -2,9 +2,12 @@
 
 struct NetworkComponent {
 	// A unique identifier for the entity on the network.
-	// A value of -1 is used to indicate a new network entity.
-	// A unique id will then be assigned by the server and clients
-	// will be notified of entity creation.
-	size_t id;
+	// A value of -1 is used to indicate an unassigned network id.
+	// A unique id will then be assigned by the server.
+	int id;
+
+	// Indicates that the associated entity has been newly created.
+	// A server should respond by notifying clients to create
+	// new ghost objects.
 	bool isNewEntity;
 };
